correctly getting css vars as strings
move NetworkVisualization to seperate component for readability
