+++ app/dashboard/_components/InterviewsTable/Columns.tsx
@@ -9,6 +9,7 @@ import { Checkbox } from '~/components/ui/checkbox';
 import { Progress } from '~/components/ui/progress';
 import TimeAgo from '~/components/ui/TimeAgo';
 import type { GetInterviewsReturnType } from '~/queries/interviews';
+import NetworkVisualization from './NetworkVisualization';
 
 export const InterviewColumns = (): ColumnDef<
   Awaited<GetInterviewsReturnType>[0]
@@ -164,145 +165,7 @@ export const InterviewColumns = (): ColumnDef<
       const network = row.original.network as NcNetwork;
       const codebook = row.original.protocol.codebook as Codebook;
 
-      if (!network || !codebook) {
-        return <div className="text-xs">No network data</div>;
-      }
-
-      // group nodes by type
-      const nodeTypeCount: Record<string, number> = {};
-      network.nodes?.forEach((node) => {
-        nodeTypeCount[node.type] = (nodeTypeCount[node.type] ?? 0) + 1;
-      });
-
-      // group edges by type
-      const edgeTypeCount: Record<string, number> = {};
-      network.edges.forEach((edge) => {
-        edgeTypeCount[edge.type] = (edgeTypeCount[edge.type] ?? 0) + 1;
-      });
-
-      if (
-        Object.keys(nodeTypeCount).length == 0 &&
-        Object.keys(edgeTypeCount).length == 0
-      ) {
-        return <div className="text-xs">No nodes or edges</div>;
-      }
-
-      return (
-        <div className="flex flex-col gap-4">
-          <div className="grid grid-cols-3 gap-8">
-            {Object.entries(nodeTypeCount).map(([nodeType, count]) => {
-              const nodeInfo = codebook.node?.[nodeType] ?? {
-                color: 'node-color-seq-1',
-                name: 'Node',
-              };
-              const color = codebook.node?.[nodeType]?.color;
-              const nodeColor = `var(--${color})`;
-              const nodeColorDark = `var(--${color}-dark)`;
-
-              console.log('node colors:', nodeColor, nodeColorDark);
-              return (
-                <div key={nodeType} className="flex flex-col items-center">
-                  <div
-                    className="flex h-8 w-8 items-center justify-center rounded-full"
-                    style={{
-                      background: `repeating-linear-gradient(
-                      145deg,
-                      var(--node-color-seq-1) 0%,
-                      var(--node-color-seq-1) 50%,
-                      var(--node-color-seq-1-dark) 50%,
-                      var(--node-color-seq-1-dark) 100%
-                      )`,
-                    }}
-                  >
-                    <span className="text-xs font-semibold text-white">
-                      {count}
-                    </span>
-                  </div>
-                  <span className="pt-1 text-xs">{nodeInfo.name}</span>
-                </div>
-              );
-            })}
-          </div>
-
-          <div className="grid grid-cols-3 gap-12">
-            {Object.entries(edgeTypeCount).map(([edgeType, count]) => {
-              const edgeInfo = codebook.edge?.[edgeType] ?? {
-                color: 'edge-color-seq-1',
-                name: 'Edge',
-              };
-              const edgeColor = `var(--${edgeInfo.color})`;
-              const edgeColorDark = `var(--${edgeInfo.color}-dark)`;
-
-              console.log(edgeColor, edgeColorDark);
-
-              return (
-                <div key={edgeType} className="flex flex-col items-center">
-                  <div className="flex items-center justify-center">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      viewBox="0 0 60 60"
-                      width="24"
-                      height="24"
-                    >
-                      <g id="Links">
-                        <circle
-                          cx="49"
-                          cy="11"
-                          r="11"
-                          fill="var(--edge-color-seq-1-dark)"
-                        />
-                        <circle
-                          cx="49"
-                          cy="49"
-                          r="11"
-                          fill="var(--edge-color-seq-1-dark)"
-                        />
-                        <circle
-                          cx="11"
-                          cy="30"
-                          r="11"
-                          fill="var(--edge-color-seq-1-dark)"
-                        />
-                        <rect
-                          x="25.3"
-                          y="20.59"
-                          width="4"
-                          height="37.64"
-                          transform="translate(-20.48 43.35) rotate(-60)"
-                          fill="var(--edge-color-seq-1-dark)"
-                        />
-                        <rect
-                          x="8.48"
-                          y="18.59"
-                          width="37.64"
-                          height="4"
-                          transform="translate(-6.64 16.41) rotate(-29.99)"
-                          fill="var(--edge-color-seq-1)"
-                        />
-                        <path
-                          d="M3.22,22.22,18.78,37.78A11,11,0,1,1,3.22,22.22Z"
-                          fill="var(--edge-color-seq-1)"
-                        />
-                        <path
-                          d="M41.22,3.22,56.78,18.78A11,11,0,1,1,41.22,3.22Z"
-                          fill="var(--edge-color-seq-1)"
-                        />
-                        <path
-                          d="M41.22,41.22,56.78,56.78A11,11,0,1,1,41.22,41.22Z"
-                          fill="var(--edge-color-seq-1)"
-                        />
-                      </g>
-                    </svg>
-                  </div>
-                  <span className="pt-1 text-xs">
-                    {edgeInfo.name} ({count})
-                  </span>
-                </div>
-              );
-            })}
-          </div>
-        </div>
-      );
+      return <NetworkVisualization network={network} codebook={codebook} />;
     },
   },
   {
